coding-agent-mcp/requirements.txt
# Core MCP framework
fastmcp>=0.1.0

# Data processing and analysis
pandas>=2.0.0
numpy>=1.24.0

# Data visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Statistical analysis
scipy>=1.10.0
scikit-learn>=1.3.0

# LangChain MCP Integration
langchain-mcp-adapters>=0.1.0
langgraph>=0.1.0
langchain>=0.2.0
langchain-openai>=0.1.0

# Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0

# Environment variables
python-dotenv>=1.0.0

# Additional utilities
pathlib2>=2.3.7; python_version < "3.4"
